+++ core/src/pb/quick/mod.rs
@@ -1,0 +1,3 @@
+// Automatically generated mod.rs
+pub mod google;
+pub mod sf;